+++ src/parser.h
@@ -33,7 +33,7 @@ bool skip_whitespace(char **src);
 bool parse_uint(char **src, uint32_t *dest, int base);
 bool parse_int(char **src, int32_t *dest, int base);
 bool parse_immediate(char **src, uint32_t *dest);
-bool parse_reg(char **src, int *index, RegisterWidth *width);
+bool parse_reg(char **src, uint8_t *index, RegisterWidth *width);
 bool parse_discrete_shift(char **src, DiscreteShift *shift);
 bool parse_immediate_shift(char **src, ShiftType *shift_type, uint8_t *shift_amount);
